Make template of homebrewware.ini
This is a file that changes every time the application is opened.
VCS tracking of this file is undesireable. To improve this situation,
the ini file is copied in the build, deploy, and distribute targets
to the root directory along with the jar file. The application currently
expects to be able to modify this file at this location.

Eventually, either a replacement should be used for the ini file, or
it should be able to be setup in preferences. A dialog asking the user
where to store this file would be useful. This could be an improvement
for all resources.